pundit installed

+++ app/controllers/items_controller.rb
@@ -3,6 +3,7 @@ class ItemsController < ApplicationController
   def create
     @user = current_user
     @item = @user.items.build(item_params)
+    authorize @item
     @new_item = Item.new
 
     if @item.save
@@ -20,7 +21,7 @@ class ItemsController < ApplicationController
   def destroy
     @user = current_user
     @item = current_user.items.find(params[:id])
-
+    authorize @item
     if @item.destroy!
       flash[:notice] = "Yey, you have completed this task ---> #{@item.name}!"
     else
